delete source aware code
