+++ apollo-federation/src/lib.rs
@@ -25,7 +25,6 @@ pub(crate) mod operation;
 pub mod query_graph;
 pub mod query_plan;
 pub mod schema;
-pub mod source_aware;
 pub mod sources;
 pub mod subgraph;
