--- apollo-federation/src/source_aware/mod.rs
@@ -1,3 +1,0 @@
-pub(crate) mod federated_query_graph;
-pub(crate) mod fetch_dependency_graph;
-pub mod query_plan;
