--- apollo-federation/src/sources/connect/query_plan/query_planner.rs
@@ -1,4 +1,0 @@
-#[derive(Debug)]
-pub struct ExecutionMetadata {
-    // pub transport: Transport,
-}
