+++ apollo-router/src/plugins/connectors/fetch.rs
@@ -5,23 +5,22 @@ use apollo_compiler::name;
 use apollo_federation::schema::ObjectFieldDefinitionPosition;
 use apollo_federation::schema::ObjectOrInterfaceFieldDefinitionPosition;
 use apollo_federation::schema::ObjectOrInterfaceFieldDirectivePosition;
-use apollo_federation::sources::connect;
 use apollo_federation::sources::connect::ConnectId;
 use apollo_federation::sources::connect::JSONSelection;
 use apollo_federation::sources::connect::SubSelection;
-use apollo_federation::sources::source;
+use apollo_federation::sources::to_remove;
 
 use crate::query_planner::fetch::FetchNode;
 use crate::query_planner::fetch::Protocol;
 use crate::query_planner::fetch::RestFetchNode;
 
-impl From<FetchNode> for source::query_plan::FetchNode {
-    fn from(value: FetchNode) -> source::query_plan::FetchNode {
+impl From<FetchNode> for to_remove::FetchNode {
+    fn from(value: FetchNode) -> to_remove::FetchNode {
         let subgraph_name = match value.protocol.as_ref() {
             Protocol::RestFetch(rf) => rf.parent_service_name.clone().into(),
             _ => value.service_name.clone(),
         };
-        source::query_plan::FetchNode::Connect(connect::query_plan::FetchNode {
+        to_remove::FetchNode::Connect(to_remove::connect::FetchNode {
             source_id: ConnectId {
                 label: value.service_name.to_string(),
                 subgraph_name,
@@ -62,7 +61,7 @@ impl FetchNode {
             connector_graph_key: connector._name(),
             parent_service_name,
         }));
-        let as_fednext_node: source::query_plan::FetchNode = self.clone().into();
+        let as_fednext_node: to_remove::FetchNode = self.clone().into();
         self.source_node = Some(Arc::new(as_fednext_node));
     }
 }
